requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pillow==10.0.1
numpy==1.25.2
easyocr==1.7.1
requests==2.31.0
groq==0.4.1
nbformat==5.9.2
pydantic==2.4.2
opencv-python-headless==4.8.0.74
python-dateutil==2.8.2